+++ src/components/Header.jsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
 
-import { AppBar, Toolbar } from "@material-ui/core";
+import { AppBar, Container, Toolbar } from "@material-ui/core";
 
 import GrowTypography from "./GrowTypography";
 
@@ -9,32 +9,69 @@ import GrowTypography from "./GrowTypography";
  * Create an AppBar/Toolbar header.
  *
  * Children are rendered inside a flexGrow Typography
- * @param   {node}    children            [description]
- * @param   {node}    [LeftAction=null]   [description]
- * @param   {Array}   [RightActions=[]]   [description]
- * @param   {Object}  [TitleProps={}}]    [description]
- * @param   {Object}  [ToolbarProps={}}]  [description]
- * @param   {any}     [rest]              [description]
+ * @param node children
+ * // TODO: Remove in v3.0.0
+ * @param node LeftAction=null
+ * @param node leftAction=null
+ * // TODO: Remove in v3.0.0
+ * @param Array RightActions=[]
+ * @param node rightAction=null
+ * @param Object TitleProps={}
+ * @param Object ToolbarProps={}
+ * @param bool withContainer if `true` wraps inner `Toolbar` inside `Container`
+ * @param any rest The rest of the props are passed to `AppBar`
  * @constructor
  */
 export default function Header({
   children,
+  // TODO: Remove in v3.0.0
   LeftAction = null,
+  leftAction = null,
+  // TODO: Remove in v3.0.0
   RightActions = [],
+  rightAction = null,
   TitleProps = {},
   ToolbarProps = {},
+  withContainer = false,
   ...rest
 }) {
+  if (LeftAction !== null) {
+    // TODO: Remove in v3.0.0
+    console.warn(
+      "Property `LeftAction` is deprecated since v2.19.2 and will be removed in v3.0.0. Use `leftAction` instead"
+    );
+  }
+  if (
+    RightActions !== null &&
+    typeof RightActions === "object" &&
+    (!RightActions.map || RightActions.length !== 0)
+  ) {
+    // TODO: Remove in v3.0.0
+    console.warn(
+      "Property `RightActions` is deprecated since v2.19.2 and will be removed in v3.0.0. Use `rightAction` instead"
+    );
+  }
+
+  const appBarContent = (
+    <Toolbar disableGutters={withContainer} {...ToolbarProps}>
+      {LeftAction}
+      {leftAction}
+      <GrowTypography variant="h6" {...TitleProps}>
+        {children}
+      </GrowTypography>
+      {RightActions}
+      {rightAction}
+    </Toolbar>
+  );
+
   return (
     <Fragment>
       <AppBar {...rest}>
-        <Toolbar {...ToolbarProps}>
-          {LeftAction}
-          <GrowTypography variant="h6" {...TitleProps}>
-            {children}
-          </GrowTypography>
-          {RightActions}
-        </Toolbar>
+        {!!withContainer ? (
+          <Container>{appBarContent}</Container>
+        ) : (
+          appBarContent
+        )}
       </AppBar>
       <Toolbar />
     </Fragment>
@@ -43,11 +80,22 @@ export default function Header({
 
 Header.propTypes = {
   children: PropTypes.node,
+  // TODO: Remove in v3.0.0
   LeftAction: PropTypes.node,
+  leftAction: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.arrayOf(PropTypes.node),
+  ]),
+  // TODO: Remove in v3.0.0
   RightActions: PropTypes.oneOfType([
     PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node)
+    PropTypes.arrayOf(PropTypes.node),
+  ]),
+  rightAction: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.arrayOf(PropTypes.node),
   ]),
   TitleProps: PropTypes.object,
-  ToolbarProps: PropTypes.object
+  ToolbarProps: PropTypes.object,
+  withContainer: PropTypes.bool,
 };