+++ src/components/index.js
@@ -4,6 +4,7 @@ export { default as BackIconButton } from "./BackIconButton";
 export { default as Collection } from "./Collection";
 export { default as Condition } from "./Condition";
 export { default as Conditional } from "./Conditional";
+export { default as ConfigWrapper } from "./ConfigWrapper";
 export { default as ConfirmDialog } from "./ConfirmDialog";
 export { default as ConfirmDialogButton } from "./ConfirmDialogButton";
 export { default as Content } from "./Content";